# docs/api/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "DeepreadQA API",
    "description": "基于 **DeepreadQA**（渐进式阅读 AgenticRAG）的 CAE 知识库问答 API。\n\n- 内置知识库：226 篇 CAE/仿真领域文档（VLM-OCR 修复语料）\n- 私有知识库：上传你自己的 markdown（collections），对其问答\n- 作答方式：BM25 检索 + 目录级渐进阅读 + rubric 对齐 compose\n- 认证：`Authorization: Bearer <key>`；错误：RFC 9457 `application/problem+json`\n- 同步/异步双模式：默认同步等待；`Prefer: respond-async` 转异步轮询\n",
    "contact": {
      "name": "DeepreadQA operator",
      "email": "support@atominfinite.ai"
    },
    "version": "1.1.0"
  },
  "paths": {
...
    "/v1/answers": {
      "post": {
        "tags": [
          "answers"
        ],
        "summary": "创建一次作答",
        "description": "对知识库提出一个问题（缺省 = 内置 CAE 库；带 `collection_id` = 你上传的私有知识库）。默认同步等待作答完成；请求头 `Prefer: respond-async` 立即返回 202 转异步轮询；`Idempotency-Key` 保证重复提交不重复计费。",
        "operationId": "create_answer_v1_answers_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/AnswerCreateRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AnswerResource"
                }
              }
            }
          },
          "202": {
            "description": "作答仍在执行（异步模式，或同步等待达到上限）；按 Location 轮询",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AnswerResource"
                }
              }
            }
          },
          "401": {
            "description": "认证失败",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              }
            }
          },
          "404": {
            "description": "资源不存在",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              }
            }
          },
          "409": {
            "description": "知识库尚无可用文档",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              }
            }
          },
          "422": {
            "description": "请求校验失败",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              }
            }
          },
          "429": {
            "description": "超出速率限制",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              }
            }
          },
          "502": {
            "description": "上游 LLM 作答失败",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              }
            }
          },
          "503": {
            "description": "服务暂不可用（未就绪或队列已满）",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              }
            }
          }
        },
        "security": [
          {
            "ApiKeyBearer": []
          }
        ]
      }
    },
    "/v1/answers/{answer_id}": {
      "get": {
        "tags": [
          "answers"
        ],
        "summary": "查询一次作答",
        "description": "返回 answer 资源当前状态。资源存在即 200（含 failed）；只有资源不存在或已过保留期才 404。",
        "operationId": "get_answer_v1_answers__answer_id__get",
        "security": [
          {
            "ApiKeyBearer": []
          }
        ],
        "parameters": [
          {
            "name": "answer_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Answer Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AnswerResource"
                }
              }
            }
          },
          "401": {
            "description": "认证失败",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              }
            }
          },
          "404": {
            "description": "资源不存在",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
      "AnswerCreateRequest": {
        "properties": {
          "question": {
            "type": "string",
            "minLength": 1,
            "title": "Question",
            "description": "面向知识库的自然语言问题（中文或英文）"
          },
          "collection_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Collection Id",
            "description": "私有知识库 id（col_…）；缺省 = 内置 CAE 知识库"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "question"
        ],
        "title": "AnswerCreateRequest",
        "description": "Body of POST /v1/answers; unknown fields are rejected.",
        "examples": [
          {
            "question": "HJC 本构模型模拟混凝土受冲击时主要考虑哪些效应？"
          }
        ]
      },
      "AnswerError": {
        "properties": {
          "code": {
            "type": "string",
            "title": "Code"
          },
          "message": {
            "type": "string",
            "title": "Message"
          }
        },
        "type": "object",
        "required": [
          "code",
          "message"
        ],
        "title": "AnswerError"
      },
      "AnswerResource": {
        "properties": {
          "id": {
            "type": "string",
            "title": "Id",
            "examples": [
              "ans_a1b2c3d4e5f60718"
            ]
          },
          "object": {
            "type": "string",
            "const": "answer",
            "title": "Object",
            "default": "answer"
          },
          "status": {
            "type": "string",
            "enum": [
              "queued",
              "running",
              "succeeded",
              "failed"
            ],
            "title": "Status"
          },
          "question": {
            "type": "string",
            "title": "Question"
          },
          "collection_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Collection Id",
            "description": "作答所用私有知识库；null = 内置 CAE 库"
          },
          "answer": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Answer",
            "description": "终答文本；succeeded 时非空"
          },
          "sources": {
            "items": {
              "$ref": "#/components/schemas/SourceDoc"
            },
            "type": "array",
            "title": "Sources",
            "description": "作答过程实际打开阅读过的文档（不含仅在检索结果中出现的候选）"
          },
          "usage": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/AnswerUsage"
              },
              {
                "type": "null"
              }
            ]
          },
          "forced_final": {
            "anyOf": [
              {
                "type": "boolean"
              },
              {
                "type": "null"
              }
            ],
            "title": "Forced Final",
            "description": "是否在轮数耗尽时被强制收敛作答"
          },
          "created_at": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Created At",
            "description": "ISO 8601 UTC"
          },
          "started_at": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Started At"
          },
          "finished_at": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Finished At"
          },
          "latency_ms": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Latency Ms",
            "description": "从开始执行到完成的耗时"
          },
          "error": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/AnswerError"
              },
              {
                "type": "null"
              }
            ],
            "description": "failed 时的错误详情"
          }
        },
        "type": "object",
        "required": [
          "id",
          "status",
          "question"
        ],
        "title": "AnswerResource",
        "description": "One answer job; the same shape is returned by POST and GET."
      },
      "AnswerUsage": {
        "properties": {
          "iterations": {
            "type": "integer",
            "title": "Iterations",
            "description": "agent 检索↔阅读回环轮数"
          },
          "total_tokens": {
            "type": "integer",
            "title": "Total Tokens",
            "description": "本次作答消耗的 LLM token 总量"
          },
          "compactions": {
            "type": "integer",
            "title": "Compactions",
            "description": "工作记忆压缩次数"
          },
          "documents_read": {
            "type": "integer",
            "title": "Documents Read",
            "description": "实际打开阅读过的文档数",
            "default": 0
          },
          "documents_seen": {
            "type": "integer",
            "title": "Documents Seen",
            "description": "检索结果中出现过的候选文档数",
            "default": 0
          }
        },
        "type": "object",
        "required": [
          "iterations",
          "total_tokens",
          "compactions"
        ],
        "title": "AnswerUsage"
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
      "Problem": {
        "properties": {
          "type": {
            "type": "string",
            "title": "Type"
          },
          "title": {
            "type": "string",
            "title": "Title"
          },
          "status": {
            "type": "integer",
            "title": "Status"
          },
          "detail": {
            "type": "string",
            "title": "Detail"
          },
          "code": {
            "type": "string",
            "title": "Code"
          },
          "request_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Request Id"
          }
        },
        "additionalProperties": true,
        "type": "object",
        "required": [
          "type",
          "title",
          "status",
          "detail",
          "code"
        ],
        "title": "Problem",
        "description": "RFC 9457 problem document; every non-2xx response uses this shape.",
        "examples": [
          {
            "code": "rate_limited",
            "detail": "API key exceeded 10 req/min",
            "request_id": "req_1a2b3c4d",
            "retry_after": 42,
            "status": 429,
            "title": "Too Many Requests",
            "type": "https://deepreadqa.dev/errors/rate-limited"
          }
        ]
      },
...
      "SourceDoc": {
        "properties": {
          "doc_id": {
            "type": "string",
            "title": "Doc Id",
            "description": "知识库内文档 id"
          },
          "title": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Title",
            "description": "文档标题"
          }
        },
        "type": "object",
        "required": [
          "doc_id"
        ],
        "title": "SourceDoc"
      },
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          },
          "input": {
            "title": "Input"
          },
          "ctx": {
            "type": "object",
            "title": "Context"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      }
    },
    "securitySchemes": {
      "ApiKeyBearer": {
        "type": "http",
        "description": "运营方签发的 API key，置于 `Authorization: Bearer <key>`",
        "scheme": "bearer"
      }
    }
  },
...
}